requirements.txt
google-generativeai
pypdf
python-dateutil
colorama
python-dotenv
streamlit
